fmap: Move package from fmap/pkg to pkg/fmap
Some tools were upset because the fmap package's name differed from the
file it was in.
